Move logic for creation product from service to JpaRepository class
